Initial changes to remove prompt agents
