--- cli/azd/extensions/azure.ai.agents/internal/pkg/agents/agent_yaml/testdata/invalid-no-model.yaml
@@ -1,4 +1,0 @@
-template:
-  kind: prompt
-  name: no-model-agent
-  instructions: Some instructions
