--- cli/azd/extensions/azure.ai.agents/internal/pkg/agents/agent_yaml/testdata/mcp-tools-agent.yaml
@@ -1,18 +1,0 @@
-template:
-  kind: prompt
-  name: mcp-tools-agent
-  description: Agent with MCP tool connections
-  model:
-    id: gpt-4o
-  tools:
-    - kind: mcp
-      name: github-mcp
-      connection:
-        kind: foundry
-        endpoint: https://api.githubcopilot.com/mcp/
-        name: github-mcp-conn
-      url: https://api.githubcopilot.com/mcp/
-    - kind: codeInterpreter
-      name: code-runner
-  instructions: |
-    You have access to GitHub via MCP and can run code.
