--- cli/azd/extensions/azure.ai.agents/internal/pkg/agents/agent_yaml/testdata/prompt-agent-minimal.yaml
@@ -1,5 +1,0 @@
-template:
-  kind: prompt
-  name: minimal-agent
-  model:
-    id: gpt-4o-mini
