--- cli/azd/extensions/azure.ai.agents/tests/samples/declarativeNoTools/agent.yaml
@@ -1,35 +1,0 @@
-template:
-  kind: prompt
-  name: Learn French Agent
-  description: |-
-    This Agent helps users learn French by providing vocabulary, grammar tips, and practice exercises.
-  metadata:
-    example:
-      - role: user
-        content: |-
-          Can you help me learn some basic French phrases for traveling?
-      - role: assistant
-        content: |- 
-          Sure! Here are some useful French phrases for your trip:
-          1. Bonjour - Hello
-          2. Merci - Thank you
-          3. Où est…? - Where is…?
-          4. Combien ça coûte? - How much does it cost?
-          5. Parlez-vous anglais? - Do you speak English?
-    tags:
-      - example
-      - learning
-    authors:
-      - jeomhove
-
-  model:
-    id: gpt-4o-mini
-    publisher: azure
-    options:
-      temperature: 0.7
-      maxTokens: 4000
-
-  instructions: |-
-    You are a helpful assistant that specializes in teaching French.
-    Provide clear explanations, examples,
-    and practice exercises to help users improve their French language skills.
